requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2023.12.12=haa95532_0
colorama=0.4.6=pypi_0
contourpy=1.2.0=pypi_0
cuda-cccl=12.3.101=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.3.101=0
cuda-runtime=11.8.0=0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.12.1=pypi_0
fonttools=4.47.0=pypi_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
imageio=2.33.1=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.4.101=0
libcurand-dev=10.3.4.101=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lightning-utilities=0.9.0=py39haa95532_0
lz4-c=1.9.4=h2bbff1b_0
matplotlib=3.8.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
networkx=3.2.1=pypi_0
numpy=1.26.2=py39h055cbcc_0
numpy-base=1.26.2=py39h65a83cf_0
opencv-python=4.8.1.78=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.12=h2bbff1b_0
packaging=23.1=py39haa95532_0
pandas=2.1.4=pypi_0
pillow=10.0.1=py39h045eedc_0
pip=23.3.1=py39haa95532_0
pyparsing=3.1.1=pypi_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pypi_0
pytorch=1.10.2=py3.9_cuda11.3_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
setuptools=68.2.2=py39haa95532_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.12.9=pypi_0
tk=8.6.12=h2bbff1b_0
torchmetrics=1.1.2=py39haa95532_0
torchvision=0.11.3=py39_cu113
tqdm=4.66.1=pypi_0
typing_extensions=4.7.1=py39haa95532_0
tzdata=2023.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py39haa95532_0
xz=5.4.5=h8cc25b3_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
